+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Internal
+
+- Added preference system
+
 ## [1.0.0-alpha.1] - 2024-10-27
 
 ### Fixed